Create library with check-out-books functionality
I should have committed ages ago. Created a library with a bookshelf. Bookshelf is a yaml file that has a list of books. A book is referenced on the bookshelf by its index. Then it can be "checked out" by setting the availability to false. It can also be checked back in.